Fix more comments
