+++ pkg/controller/master/migration/vmim_controller_test.go
@@ -842,3 +842,160 @@ func TestHandler_OnVmimChanged_WithResourceQuota_WithCompensation(t *testing.T) 
 		})
 	}
 }
+
+func Test_isVmiResetHarvesterMigrationAnnotationRequired(t *testing.T) {
+	type args struct {
+		vmi *kubevirtv1.VirtualMachineInstance
+	}
+
+	tests := []struct {
+		name string
+		args args
+		want bool
+	}{
+		{
+			name: "vmi migration is ongoing, does not require clean",
+			args: args{
+				vmi: &kubevirtv1.VirtualMachineInstance{
+					ObjectMeta: v1.ObjectMeta{
+						Name:      "vm1",
+						Namespace: resourceQuotaNamespace,
+						UID:       uid,
+						Annotations: map[string]string{
+							util.AnnotationMigrationUID:   vmimUID,
+							util.AnnotationMigrationState: StatePending,
+						},
+					},
+					Status: kubevirtv1.VirtualMachineInstanceStatus{
+						MigrationState: &kubevirtv1.VirtualMachineInstanceMigrationState{
+							Completed:    false,
+							MigrationUID: vmimUID,
+						},
+					},
+				},
+			},
+			want: false,
+		},
+		{
+			name: "vmi migration is completed, require clean",
+			args: args{
+				vmi: &kubevirtv1.VirtualMachineInstance{
+					ObjectMeta: v1.ObjectMeta{
+						Name:      "vm1",
+						Namespace: resourceQuotaNamespace,
+						UID:       uid,
+						Annotations: map[string]string{
+							util.AnnotationMigrationUID:   vmimUID,
+							util.AnnotationMigrationState: StatePending,
+						},
+					},
+					Status: kubevirtv1.VirtualMachineInstanceStatus{
+						MigrationState: &kubevirtv1.VirtualMachineInstanceMigrationState{
+							Completed:    true,
+							MigrationUID: vmimUID,
+						},
+					},
+				},
+			},
+			want: true,
+		},
+		{
+			name: "vmi migration is completed, require clean",
+			args: args{
+				vmi: &kubevirtv1.VirtualMachineInstance{
+					ObjectMeta: v1.ObjectMeta{
+						Name:      "vm1",
+						Namespace: resourceQuotaNamespace,
+						UID:       uid,
+						Annotations: map[string]string{
+							util.AnnotationMigrationUID:   vmimUID,
+							util.AnnotationMigrationState: StatePending,
+						},
+					},
+					Status: kubevirtv1.VirtualMachineInstanceStatus{
+						MigrationState: &kubevirtv1.VirtualMachineInstanceMigrationState{
+							Failed:       true,
+							MigrationUID: vmimUID,
+						},
+					},
+				},
+			},
+			want: true,
+		},
+		{
+			name: "vmi migration is completed but MigrationUID does not match, does not require clean",
+			args: args{
+				vmi: &kubevirtv1.VirtualMachineInstance{
+					ObjectMeta: v1.ObjectMeta{
+						Name:      "vm1",
+						Namespace: resourceQuotaNamespace,
+						UID:       uid,
+						Annotations: map[string]string{
+							util.AnnotationMigrationUID:   "unrelated",
+							util.AnnotationMigrationState: StatePending,
+						},
+					},
+					Status: kubevirtv1.VirtualMachineInstanceStatus{
+						MigrationState: &kubevirtv1.VirtualMachineInstanceMigrationState{
+							Failed:       true,
+							MigrationUID: vmimUID,
+						},
+					},
+				},
+			},
+			want: false,
+		},
+		{
+			name: "vmi migration is completed, already cleaned",
+			args: args{
+				vmi: &kubevirtv1.VirtualMachineInstance{
+					ObjectMeta: v1.ObjectMeta{
+						Name:        "vm1",
+						Namespace:   resourceQuotaNamespace,
+						UID:         uid,
+						Annotations: map[string]string{
+							// util.AnnotationMigrationUID:   vmimUID,
+							// util.AnnotationMigrationState: StatePending,
+						},
+					},
+					Status: kubevirtv1.VirtualMachineInstanceStatus{
+						MigrationState: &kubevirtv1.VirtualMachineInstanceMigrationState{
+							Completed:    true,
+							MigrationUID: vmimUID,
+						},
+					},
+				},
+			},
+			want: false,
+		},
+		{
+			name: "vmi migration is failed, already cleaned",
+			args: args{
+				vmi: &kubevirtv1.VirtualMachineInstance{
+					ObjectMeta: v1.ObjectMeta{
+						Name:        "vm1",
+						Namespace:   resourceQuotaNamespace,
+						UID:         uid,
+						Annotations: map[string]string{
+							// util.AnnotationMigrationUID:   vmimUID,
+							// util.AnnotationMigrationState: StatePending,
+						},
+					},
+					Status: kubevirtv1.VirtualMachineInstanceStatus{
+						MigrationState: &kubevirtv1.VirtualMachineInstanceMigrationState{
+							Failed:       true,
+							MigrationUID: vmimUID,
+						},
+					},
+				},
+			},
+			want: false,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			resetRequired := isVmiResetHarvesterMigrationAnnotationRequired(tt.args.vmi)
+			assert.Equal(t, tt.want, resetRequired, "case %q", tt.name)
+		})
+	}
+}
